refactor: when switching educations tabs switch the url as well
