--- src/app/(pages)/educations/_components/Evidence/EducationsEvidenceTab.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const EducationsEvidenceTab: React.FC = () => {
-  return (
-    <div>
-      <h2>Tab 2 content</h2>
-      {/* Add your table content here */}
-    </div>
-  );
-};
-
-export default EducationsEvidenceTab;
