+++ src/app/(pages)/educations/evidence/page.tsx
@@ -1,0 +1,10 @@
+const EducationsEvidencePage = () => {
+  return (
+    <div>
+      <h2>Tab 2 content</h2>
+      {/* Add your table content here */}
+    </div>
+  );
+};
+
+export default EducationsEvidencePage;
