+++ src/app/(pages)/educations/list/EducationsListSearch.tsx
