+++ src/app/(pages)/educations/list/EducationsListTable.tsx
@@ -7,7 +7,7 @@ import {
   TableRow,
 } from "~/components/organisms/Table";
 import PaginationComponent from "~/components/organisms/pagination/PaginationComponent";
-import { useState, useEffect, type FC } from "react";
+import { useState, type FC } from "react";
 import type { FindEducationReturnDTO } from "~/server/services/education/education.repository";
 import { translateEducationType } from "~/app/(pages)/educations/utils";
 import { type EducationType } from "~/server/db/schema";
@@ -82,7 +82,7 @@ const EducationsListTable: FC<Props> = ({ data, totalPageNumber, refetch }) => {
                 <TableCell className="cursor-pointer md:table-cell">
                   <Link
                     href={{
-                      pathname: `/educations/${license.id}`,
+                      pathname: `/educations/list/${license.id}`,
                     }}
                   >
                     <Pencil />
