+++ src/app/(pages)/educations/list/[id]/_components/EducationsForm.tsx
