+++ src/components/atoms/Tabs.tsx
@@ -1,4 +1,4 @@
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { usePathname, useRouter } from "next/navigation";
 import React, { type FC, useEffect, useState } from "react";
 import { cn } from "~/components/utils";
 
@@ -8,31 +8,16 @@ type Props = {
 
 export type TabProp = {
   label: string;
-  queryParam: string;
-  content: React.ReactNode;
+  link: string;
 };
 
 const Tabs: FC<Props> = ({ tabs }: Props) => {
   const router = useRouter();
-  const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const [activeTab, setActiveTab] = useState<string>(tabs[0]!.queryParam);
+  const [activeTab, setActiveTab] = useState<string>(tabs[0]!.link);
 
   useEffect(() => {
-    const current = new URLSearchParams(Array.from(searchParams.entries())); // -> has to use this form
-
-    if (activeTab === undefined) {
-      current.delete("selected");
-    } else {
-      current.set("selected", activeTab.toString());
-    }
-
-    // cast to string
-    const search = current.toString();
-    const query = search ? `?${search}` : "";
-
-    router.push(`${pathname}${query}`);
-  }, [activeTab, pathname, router, searchParams]);
+    router.push(`/educations/${activeTab}`);
+  }, [activeTab, router]);
 
   if (!tabs || tabs.length === 0) {
     return <div className="text-center text-gray-500">No tabs provided</div>;
@@ -44,26 +29,20 @@ const Tabs: FC<Props> = ({ tabs }: Props) => {
         {tabs.map((tab, index) => (
           <button
             key={index}
-            onClick={() => setActiveTab(tab.queryParam)}
+            onClick={() => setActiveTab(tab.link)}
             className={cn(
               "cursor-pointer rounded-lg py-2 text-sm font-medium",
               "focus:outline-none focus:ring-1 focus:ring-red-600",
-              activeTab === tab.queryParam
+              activeTab === tab.link
                 ? "border-red px-1 text-red-600"
                 : "hover:bg-red-100",
             )}
-            aria-selected={activeTab === tab.queryParam}
+            aria-selected={activeTab === tab.link}
             role="tab"
           >
             {tab.label}
           </button>
         ))}
-      </div>
-      <div className="mt-4">
-        {
-          tabs[tabs.findIndex(({ queryParam }) => queryParam === activeTab)]
-            ?.content
-        }
       </div>
     </div>
   );
